Fix test build failure on Windows
